+++ Changelog.md
@@ -1,3 +1,11 @@
+# 0.4.0
+
+## ElectronNET.Core
+
+- Fixed ElectronSingleInstance handling (#996) @softworkz
+- Fixed `PackageId` handling (#993) @softworkz
+- Added cross-platform npm restore and check mismatch on publish (#988) @softworkz
+
 # 0.3.1
 
 ## ElectronNET.Core